+++ etc/blocks/monetization-test/workflow/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @milaboratories/milaboratories.monetization-test.workflow
 
+## 1.0.6
+
+### Patch Changes
+
+- 14b832c: fix mnz-client in edge cases like empty arg list, empty runs
+
+## 1.0.5
+
+### Patch Changes
+
+- 5828a76: fix mnz-client in edge cases like empty arg list, empty runs
+
 ## 1.0.4
 
 ### Patch Changes